requirements.txt
Flask==2.2.2
gunicorn==20.1.0
Jinja2==3.1.2
jsonschema==4.17.3
PyYAML==6.0
requests==2.28.2
watchdog==2.2.1
Werkzeug==2.2.2

